sessions: Add speaker data set
This adds a separate data set for the speakers. This allows us to define
attributes which are common to each speaker, such as their id or bio.

At the moment, it is required that each speaker has a unique public
name, as we rely on this to merge the session data with the speaker data
to generate the schedule.json.

Signed-off-by: Sebastian Wicki <gandro@gmx.net>